# backend/public/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "version": "1.0.0",
        "title": "New Life Nursery",
        "description": "New Life Nursery is a wholesale plant distributor, located in Southern New Jersey. We strive to provide beautiful, healthy plants at competitive prices.",
        "termsOfService": "https://newlifenurseryinc.com/terms-and-conditions",
        "license": {
            "name": "MIT"
        },
        "contact": {
            "email": "newlifenurseryinc@gmail.com"
        }
    },
    "servers": [
        {
            "url": "http://newlifenurseryinc.com/v1"
        }
    ],
    "paths": {
...
        "/profile": {
            "get": {
                "summary": "User's profile information",
                "operationId": "getProfile",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/Session"
                    },
                    {
                        "$ref": "#/components/parameters/UserId"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Details about all customers",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Customer"
                                }
                            }
                        }
                    },
                    "default": {
                        "$ref": "#/components/responses/UnexpectedError"
                    }
                }
            },
            "put": {
                "summary": "Update user's profile",
                "operationId": "updateProfile",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/Session"
                    },
                    {
                        "$ref": "#/components/parameters/UserId"
                    },
                    {
                        "$ref": "#/components/parameters/Customer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "default": {
                        "$ref": "#/components/responses/UnexpectedError"
                    }
                }
            }
        },
...
    },
    "components": {
        "parameters": {
...
            "UserId": {
                "name": "userId",
                "in": "header",
                "description": "a user's uuid. Used by admins to retrieve data of customers",
                "required": true,
                "schema": {
                    "type": "string"
                }
            },
            "Customer": {
                "name": "customerData",
                "in": "header",
                "description": "customer data, including cart",
                "required": true,
                "schema": {
                    "type": "#/components/schemas/Session"
                }
            },
            "Session": {
                "name": "session",
                "in": "cookie",
                "description": "current user session",
                "required": true,
                "schema": {
                    "$ref": "#/components/schemas/Session"
                }
            },
...
        },
        "schemas": {
            "Session": {
                "type": "object",
                "required": [
                    "tag",
                    "token"
                ],
                "properties": {
                    "tag": {
                        "type": "string"
                    },
                    "token": {
                        "type": "string"
                    }
                }
            },
...
            "Customer": {
                "type": "object",
                "required": [
                    "id"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "tag": {
                        "type": "string"
                    },
                    "first_name": {
                        "type": "string"
                    },
                    "last_name": {
                        "type": "string"
                    },
                    "pronouns": {
                        "type": "string"
                    },
                    "business": {
                        "type": "string"
                    },
                    "theme": {
                        "type": "string"
                    },
                    "existing_customer": {
                        "type": "boolean"
                    },
                    "roles": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Role"
                        }
                    },
                    "emails": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Email"
                        }
                    },
                    "phones": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Phone"
                        }
                    },
                    "cart_id": {
                        "type": "integer"
                    },
                    "cart_size": {
                        "type": "integer"
                    }
                }
            },
...
            "Role": {
                "type": "object",
                "properties": {
                    "title": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    }
                }
            },
            "Email": {
                "type": "object",
                "properties": {
                    "email_address": {
                        "type": "string",
                        "example": "somerandomemail@gmail.com"
                    },
                    "receives_delivery_updates": {
                        "type": "boolean",
                        "example": true
                    }
                }
            },
            "Phone": {
                "type": "object",
                "required": [
                    "unformatted_number"
                ],
                "properties": {
                    "unformatted_number": {
                        "type": "string"
                    },
                    "country_code": {
                        "type": "string"
                    },
                    "extension": {
                        "type": "string"
                    },
                    "is_mobile": {
                        "type": "boolean"
                    },
                    "receives_delivery_updates": {
                        "type": "boolean"
                    }
                }
            },
...
            "Error": {
                "type": "object",
                "required": [
                    "code",
                    "message"
                ],
                "properties": {
                    "code": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "msg": {
                        "type": "string"
                    }
                }
            }
        },
        "responses": {
            "UnexpectedError": {
                "description": "unexpected error",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Error"
                        }
                    }
                }
            }
        }
    }
}